requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.1
cryptography==39.0.1
Django==3.2.18
django-cors-headers==3.13.0
djangorestframework==3.14.0
http-ece==1.1.0
idna==3.4
importlib-metadata==4.13.0
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
py-vapid==1.9.0
pycparser==2.21
python-dotenv==0.21.1
pytz==2022.7.1
pywebpush==1.14.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
watchdog==2.3.1
wcwidth==0.2.6
zipp==3.15.0
